Fix issue when searching and sorting in list.js
By default, list.js uses the entire contents of a tag to perform
comparisons with. However, this includes the tags/attributes of any
children and this can interfere with the expected results. By stripping
these out with a regex, the search and sort functions work much better.

This said, the sort desc function is not quite perfect - numeric strings
always sit at the top regardless of whether the sort is by asc or desc.